Fixed concatenation.

+++ trainers/ss/ss_gan_trainer.py
@@ -21,7 +21,7 @@ class SSGanTrainer(GanTrainer):
 
     def get_ss_loss(self):
         alpha = 0.9
-        real_labels = tf.concat([self.labels, tf.zeros([self.batch_size, 1])], axis=1)
+        real_labels = tf.concat([self.labels, tf.zeros([self.batch_size, 1], dtype=tf.int32)], axis=1)
         fake_labels = tf.concat([(1-alpha)*tf.ones([self.batch_size, self.classes_num])/self.classes_num,
                                  alpha*tf.ones([self.batch_size, 1])], axis=1)
         inv_fake_labels = tf.concat([alpha*tf.ones([self.batch_size, self.classes_num]),
